Extract editor player start helpers
