+++ CMakeLists.txt
@@ -159,6 +159,7 @@ target_sources(
         src/game_data_conditions_presentation.c
         src/game_data_controller_runtime.c
         src/game_data_controller_binding_helpers.c
+        src/game_data_editor_player_start.c
         src/game_data_grid_actions.c
         src/game_data_grid_runtime.c
         src/game_data_input_device_helpers.c
