+++ src/game_data_internal.h
@@ -1043,6 +1043,8 @@ int sector_level_find_sector_name(const sector_level_runtime *level, const char 
 const brush_world_runtime *find_brush_world_runtime(const slayer3d_game_data_runtime *runtime, const char *name);
 brush_world_runtime *find_brush_world_runtime_mutable(slayer3d_game_data_runtime *runtime, const char *name);
 void editor_brush_world_mark_dirty(brush_world_runtime *world_runtime);
+const editor_player_start_runtime *find_editor_player_start(const slayer3d_game_data_runtime *runtime,
+                                                            const char *name);
 bool compile_brush_world_visibility_grid(brush_world_runtime *world_runtime);
 void free_brush_world_visibility_grid(brush_world_runtime *world_runtime);
 bool rebuild_brush_world_runtime_artifacts(brush_world_runtime *world_runtime, char *error_buffer,
